Make the dotfiles docker run as the current user
Dockerhub currently builds the base dotfiles docker. As such,
it does not know about the user who is going to be running the docker
and ends up running as root. One issue this was causing is files
created within the dotfiles docker were owned by the docker's root
user rather than the user running the docker.

In this change we make use the dotfiles docker as a base for a user-
specific dotfiles docker. We pass in the user info into the docker build,
and symlink the user directory in the docker to the root directory
where the dotfiles were originally installed.